fix crontab environment variables
